backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
openai>=1.50.0
pillow>=10.4.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
httpx>=0.27.0
replicate>=0.25.0
numpy>=1.26.0
scipy>=1.12.0
opencv-python>=4.9.0
tenacity>=9.0.0

